multi_client/requirements.txt
grpcio
grpcio-tools
pyaudio